+++ src/command.rs
@@ -33,3 +33,51 @@ pub struct Argument {
     /// The argument value. Can be empty.
     pub value: String,
 }
+
+impl Argument {
+    /// Parses the argument value as a T using core::str::parse(). Convenience
+    /// method that returns a string error to easy error handling in a
+    /// [`Runner`](crate::Runner).
+    pub fn parse<T>(&self) -> Result<T, String>
+    where
+        T: std::str::FromStr,
+        <T as std::str::FromStr>::Err: std::fmt::Display,
+    {
+        self.value.parse().map_err(|e| format!("invalid argument '{}': {e}", self.value))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    /// Basic tests of Argument.parse(). Not comprehensive, since it dispatches
+    /// to core::str::parse().
+    #[test]
+    fn argument_parse() {
+        macro_rules! arg {
+            ($s:expr) => {
+                Argument { key: None, value: $s.to_string() }
+            };
+        }
+
+        assert_eq!(arg!("-1").parse(), Ok(-1_i64));
+        assert_eq!(arg!("0").parse(), Ok(0_i64));
+        assert_eq!(arg!("1").parse(), Ok(1_i64));
+        assert_eq!(
+            arg!("").parse::<i64>(),
+            Err("invalid argument '': cannot parse integer from empty string".to_string())
+        );
+        assert_eq!(
+            arg!("foo").parse::<i64>(),
+            Err("invalid argument 'foo': invalid digit found in string".to_string())
+        );
+
+        assert_eq!(arg!("false").parse(), Ok(false));
+        assert_eq!(arg!("true").parse(), Ok(true));
+        assert_eq!(
+            arg!("").parse::<bool>(),
+            Err("invalid argument '': provided string was not `true` or `false`".to_string())
+        );
+    }
+}